requirements.txt
# Core API framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
websockets==13.1

# LLM Orchestration
langchain==0.3.7
langgraph==0.2.39
langchain-groq==0.2.1
langchain-ollama==0.2.0
langsmith==0.1.147

# Speech-to-Text (Groq Whisper)
groq==0.12.0

# Text-to-Speech (Edge-TTS, no API key)
edge-tts

# Sentiment Analysis (HuggingFace)
requests==2.32.3

# Database (SQLite ORM)
sqlalchemy==2.0.36

# Audio processing
pydub==0.25.1

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
aiofiles==24.1.0
httpx==0.27.2
